FastAPI Blog Platform with JWT/requirements.txt
anyio==3.7.0
psycopg2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
python-jose==3.3.0
python-multipart
rsa==4.9
six==1.16.0
sniffio==1.3.0
databases
gino
SQLAlchemy
starlette==0.27.0
typing-extensions==4.6.2
uvicorn==0.22.0
